linked database with java code

+++ project/src/main/java/com/sshgroup/ssh_fridge_contents_tracker/Ingredients.java
@@ -12,8 +12,7 @@ import java.util.Date;
 @Table(name = "ingredients_table")
 public class Ingredients {
     @Id
-    int ingredients_id;
-
+    long ingredients_id;
 
     @Column
     private String ingredients_name;
@@ -27,14 +26,14 @@ public class Ingredients {
 
     }
 
-    public Ingredients( int ingredients_id, String ingredients, int quantity_available, int cost_per_kg) {
+    public Ingredients( long ingredients_id, String ingredients, int quantity_available, int cost_per_kg) {
 
         this.ingredients_id = ingredients_id;
         this.quantity_available = quantity_available;
         this.cost_per_kg = cost_per_kg;
 
     }
-    public int getIngredients_id() {
+    public long getIngredients_id() {
         return ingredients_id;
     }
